Finalize updates for server launch
Styles made consistent, and a few unexpected responsive flukes adjusted.

+++ flask_app/models/favorite.py
@@ -14,7 +14,7 @@ class Favorite():
     
     @classmethod
     def get_favorites(cls, data):
-        query = "SELECT recipe_id FROM favorites WHERE user_id = %(user_id)s AND recipe_id = %(recipe_id)s;"
+        query = "SELECT recipe_id FROM favorites WHERE user_id = %(user_id)s;"
         results = connectToMySQL(DATABASE).query_db(query, data)
         if results == ():
             return False
@@ -23,5 +23,5 @@ class Favorite():
     
     @classmethod
     def remove_piq(cls, data):
-        query = "DELETE FROM favorites WHERE recipe_id = %(recipe_id)s;"
+        query = "DELETE FROM favorites WHERE user_id = %(user_id)s AND recipe_id = %(recipe_id)s;"
         return connectToMySQL(DATABASE).query_db(query, data)
